+++ bitalarm-manager-node/src/main/java/com/dwarfeng/bitalarmmanager/node/controller/v1/AlarmInfoController.java
@@ -1,0 +1,118 @@
+package com.dwarfeng.bitalarmmanager.node.controller.v1;
+
+import com.dwarfeng.bitalarm.sdk.bean.entity.JSFixedFastJsonAlarmInfo;
+import com.dwarfeng.bitalarm.sdk.bean.entity.WebInputAlarmInfo;
+import com.dwarfeng.bitalarm.stack.bean.entity.AlarmInfo;
+import com.dwarfeng.bitalarmmanager.stack.service.AlarmInfoResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.bean.key.JSFixedFastJsonLongIdKey;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
+import com.dwarfeng.subgrade.sdk.validation.group.Insert;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 报警信息控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+@RestController
+@RequestMapping("/api/v1")
+public class AlarmInfoController {
+
+    @Autowired
+    private AlarmInfoResponseService service;
+    @Autowired
+    private ServiceExceptionMapper sem;
+
+    @Autowired
+    private BeanTransformer<AlarmInfo, JSFixedFastJsonAlarmInfo> beanTransformer;
+
+    @GetMapping("/alarm-info/{id}/exists")
+    @BehaviorAnalyse
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            boolean exists = service.exists(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Boolean.class, e, sem));
+        }
+    }
+
+    @GetMapping("/alarm-info/{id}")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonAlarmInfo> get(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            AlarmInfo alarmInfo = service.get(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonAlarmInfo.of(alarmInfo)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonAlarmInfo.class, e, sem));
+        }
+    }
+
+    @PostMapping("/alarm-info")
+    @BehaviorAnalyse
+    @BindingCheck
+    public FastJsonResponseData<JSFixedFastJsonLongIdKey> insert(
+            HttpServletRequest request,
+            @RequestBody @Validated(Insert.class) WebInputAlarmInfo alarmInfo, BindingResult bindingResult) {
+        try {
+            LongIdKey insert = service.insert(WebInputAlarmInfo.toStackBean(alarmInfo));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonLongIdKey.of(insert)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonLongIdKey.class, e, sem));
+        }
+    }
+
+    @PatchMapping("/alarm-info")
+    @BehaviorAnalyse
+    @BindingCheck
+    public FastJsonResponseData<Object> update(
+            HttpServletRequest request,
+            @RequestBody @Validated WebInputAlarmInfo alarmInfo, BindingResult bindingResult) {
+        try {
+            service.update(WebInputAlarmInfo.toStackBean(alarmInfo));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @DeleteMapping("/alarm-info/{id}")
+    @BehaviorAnalyse
+    public FastJsonResponseData<Object> delete(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            service.delete(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @GetMapping("/alarm-info/all")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonAlarmInfo>> all(
+            HttpServletRequest request, @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            PagedData<AlarmInfo> all = service.all(new PagingInfo(page, rows));
+            PagedData<JSFixedFastJsonAlarmInfo> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+}