+++ bitalarm-manager-node/src/main/java/com/dwarfeng/bitalarmmanager/node/controller/v1/AlarmSettingController.java
@@ -1,0 +1,160 @@
+package com.dwarfeng.bitalarmmanager.node.controller.v1;
+
+import com.dwarfeng.bitalarm.sdk.bean.entity.JSFixedFastJsonAlarmSetting;
+import com.dwarfeng.bitalarm.sdk.bean.entity.WebInputAlarmSetting;
+import com.dwarfeng.bitalarm.stack.bean.entity.AlarmSetting;
+import com.dwarfeng.bitalarmmanager.stack.service.AlarmSettingResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.bean.key.JSFixedFastJsonLongIdKey;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
+import com.dwarfeng.subgrade.sdk.validation.group.Insert;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 报警设置控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+@RestController
+@RequestMapping("/api/v1")
+public class AlarmSettingController {
+
+    @Autowired
+    private AlarmSettingResponseService service;
+    @Autowired
+    private ServiceExceptionMapper sem;
+
+    @Autowired
+    private BeanTransformer<AlarmSetting, JSFixedFastJsonAlarmSetting> beanTransformer;
+
+    @GetMapping("/alarm-setting/{id}/exists")
+    @BehaviorAnalyse
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            boolean exists = service.exists(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Boolean.class, e, sem));
+        }
+    }
+
+    @GetMapping("/alarm-setting/{id}")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonAlarmSetting> get(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            AlarmSetting alarmSetting = service.get(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonAlarmSetting.of(alarmSetting)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonAlarmSetting.class, e, sem));
+        }
+    }
+
+    @PostMapping("/alarm-setting")
+    @BehaviorAnalyse
+    @BindingCheck
+    public FastJsonResponseData<JSFixedFastJsonLongIdKey> insert(
+            HttpServletRequest request,
+            @RequestBody @Validated(Insert.class) WebInputAlarmSetting alarmSetting, BindingResult bindingResult) {
+        try {
+            LongIdKey insert = service.insert(WebInputAlarmSetting.toStackBean(alarmSetting));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonLongIdKey.of(insert)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonLongIdKey.class, e, sem));
+        }
+    }
+
+    @PatchMapping("/alarm-setting")
+    @BehaviorAnalyse
+    @BindingCheck
+    public FastJsonResponseData<Object> update(
+            HttpServletRequest request,
+            @RequestBody @Validated WebInputAlarmSetting alarmSetting, BindingResult bindingResult) {
+        try {
+            service.update(WebInputAlarmSetting.toStackBean(alarmSetting));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @DeleteMapping("/alarm-setting/{id}")
+    @BehaviorAnalyse
+    public FastJsonResponseData<Object> delete(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            service.delete(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @GetMapping("/alarm-setting/all")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonAlarmSetting>> all(
+            HttpServletRequest request, @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            PagedData<AlarmSetting> all = service.all(new PagingInfo(page, rows));
+            PagedData<JSFixedFastJsonAlarmSetting> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("/point/{pointId}/alarm-setting")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonAlarmSetting>> childForPoint(
+            HttpServletRequest request,
+            @PathVariable("pointId") long pointId, @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            PagedData<AlarmSetting> all = service.childForPoint(pointId, new PagingInfo(page, rows));
+            PagedData<JSFixedFastJsonAlarmSetting> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("/alarm-setting/alarm-message-like")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonAlarmSetting>> alarmMessageLike(
+            HttpServletRequest request,
+            @RequestParam("pattern") String pattern, @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            PagedData<AlarmSetting> all = service.alarmMessageLike(pattern, new PagingInfo(page, rows));
+            PagedData<JSFixedFastJsonAlarmSetting> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+
+    @GetMapping("/alarm-setting/alarm-type-equals")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonAlarmSetting>> alarmTypeEquals(
+            HttpServletRequest request,
+            @RequestParam("type") byte type, @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            PagedData<AlarmSetting> all = service.alarmTypeEquals(type, new PagingInfo(page, rows));
+            PagedData<JSFixedFastJsonAlarmSetting> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+}