+++ bitalarm-manager-node/src/main/java/com/dwarfeng/bitalarmmanager/node/controller/v1/CurrentAlarmController.java
@@ -1,0 +1,118 @@
+package com.dwarfeng.bitalarmmanager.node.controller.v1;
+
+import com.dwarfeng.bitalarm.sdk.bean.entity.JSFixedFastJsonCurrentAlarm;
+import com.dwarfeng.bitalarm.sdk.bean.entity.WebInputCurrentAlarm;
+import com.dwarfeng.bitalarm.stack.bean.entity.CurrentAlarm;
+import com.dwarfeng.bitalarmmanager.stack.service.CurrentAlarmResponseService;
+import com.dwarfeng.subgrade.sdk.bean.dto.FastJsonResponseData;
+import com.dwarfeng.subgrade.sdk.bean.dto.JSFixedFastJsonPagedData;
+import com.dwarfeng.subgrade.sdk.bean.dto.PagingUtil;
+import com.dwarfeng.subgrade.sdk.bean.dto.ResponseDataUtil;
+import com.dwarfeng.subgrade.sdk.bean.key.JSFixedFastJsonLongIdKey;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.http.BindingCheck;
+import com.dwarfeng.subgrade.sdk.validation.group.Insert;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * 当前报警控制器。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+@RestController
+@RequestMapping("/api/v1")
+public class CurrentAlarmController {
+
+    @Autowired
+    private CurrentAlarmResponseService service;
+    @Autowired
+    private ServiceExceptionMapper sem;
+
+    @Autowired
+    private BeanTransformer<CurrentAlarm, JSFixedFastJsonCurrentAlarm> beanTransformer;
+
+    @GetMapping("/current-alarm/{id}/exists")
+    @BehaviorAnalyse
+    public FastJsonResponseData<Boolean> exists(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            boolean exists = service.exists(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(exists));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Boolean.class, e, sem));
+        }
+    }
+
+    @GetMapping("/current-alarm/{id}")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonCurrentAlarm> get(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            CurrentAlarm currentAlarm = service.get(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonCurrentAlarm.of(currentAlarm)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonCurrentAlarm.class, e, sem));
+        }
+    }
+
+    @PostMapping("/current-alarm")
+    @BehaviorAnalyse
+    @BindingCheck
+    public FastJsonResponseData<JSFixedFastJsonLongIdKey> insert(
+            HttpServletRequest request,
+            @RequestBody @Validated(Insert.class) WebInputCurrentAlarm currentAlarm, BindingResult bindingResult) {
+        try {
+            LongIdKey insert = service.insert(WebInputCurrentAlarm.toStackBean(currentAlarm));
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonLongIdKey.of(insert)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonLongIdKey.class, e, sem));
+        }
+    }
+
+    @PatchMapping("/current-alarm")
+    @BehaviorAnalyse
+    @BindingCheck
+    public FastJsonResponseData<Object> update(
+            HttpServletRequest request,
+            @RequestBody @Validated WebInputCurrentAlarm currentAlarm, BindingResult bindingResult) {
+        try {
+            service.update(WebInputCurrentAlarm.toStackBean(currentAlarm));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @DeleteMapping("/current-alarm/{id}")
+    @BehaviorAnalyse
+    public FastJsonResponseData<Object> delete(HttpServletRequest request, @PathVariable("id") long id) {
+        try {
+            service.delete(new LongIdKey(id));
+            return FastJsonResponseData.of(ResponseDataUtil.good(null));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(Object.class, e, sem));
+        }
+    }
+
+    @GetMapping("/current-alarm/all")
+    @BehaviorAnalyse
+    public FastJsonResponseData<JSFixedFastJsonPagedData<JSFixedFastJsonCurrentAlarm>> all(
+            HttpServletRequest request, @RequestParam("page") int page, @RequestParam("rows") int rows) {
+        try {
+            PagedData<CurrentAlarm> all = service.all(new PagingInfo(page, rows));
+            PagedData<JSFixedFastJsonCurrentAlarm> transform = PagingUtil.transform(all, beanTransformer);
+            return FastJsonResponseData.of(ResponseDataUtil.good(JSFixedFastJsonPagedData.of(transform)));
+        } catch (Exception e) {
+            return FastJsonResponseData.of(ResponseDataUtil.bad(JSFixedFastJsonPagedData.class, e, sem));
+        }
+    }
+}